# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: ComConnect API
  version: 2.0.0
  description: Public API gateway for workspaces, real-time chat, tasks, notifications, and AI agents.
servers:
  - url: /
    description: Current deployment
...
paths:
  /api/user:
    post:
      tags: [Identity]
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
      responses:
        "201":
          description: User created
    get:
      tags: [Identity]
      summary: Search users
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Users
...
  /api/workspace:
    post:
      tags: [Workspaces]
      summary: Create a workspace
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Workspace created
  /api/workspace/user:
    get:
      tags: [Workspaces]
      summary: List the current user's workspaces
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Workspaces
...
  /api/message:
    post:
      tags: [Chat]
      summary: Enqueue and persist a message through Redis Streams
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatId, content]
              properties:
                chatId: { type: string }
                content: { type: string }
      responses:
        "201":
          description: Message persisted by the message worker
        "503":
          description: Message stream or persistence worker unavailable
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    UserCredentials:
      type: object
      required: [email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }